oe.headrevs: save and restore the bitbake URI headrevs
This module provides functions to save and restore the BB_URI_HEADREVS from/to
the persist_data database. The BB_URI_HEADREVS are the mappings between git
refs and the revisions they refer to.

This is useful to fully support BB_NO_NETWORK when AUTOREV is involved, as we
can use the dumped mappings rather than contacting upstream at parse time.

Signed-off-by: Christopher Larson <chris_larson@mentor.com>