lirc: correct python deps from 2 to 3
Lirc requires python 3 for its build tools, and also requires python 3 for its
scripts on target. ${PN}-python is also added to RRECOMMENDS_${PN}, as this
dependency exists regardless. The main lirc package already required this
package and python itself, as the lirc package includes irdb-get and
pronto2lirc, which are python 3 scripts, and lirc-setup, which is a symlink to
a python module in the lirc-python package.

Signed-off-by: Christopher Larson <chris_larson@mentor.com>